pyproject.toml
[project]
name = "e-treasure-hunt"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "channels>=4",
  "daphne>=4",
  "django>=5",
  "djangorestframework>=3.15",
  "geopy>=2.4",
  "holidays>=0.54",
  "pillow>=12",
  "whitenoise>=6.4",
]

[project.optional-dependencies]
azure = [
  "azure-identity>=1.13",
  "channels-redis>=4.1",
  "django-storages[azure]>=1.13.1",
  "mssql-django>=1.4",
  "redis-entraid>=1",
]

[dependency-groups]
dev = [
  "aiodns>=3",
  "aiohttp>=3.9",
  "django-stubs>=5",
  "djangorestframework-stubs>=1.8",
  "matplotlib>=3.10.6",
  "mypy>=1.11.1",
  "pydantic>=2.4",
  "pytest>=8",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "pytest-django>=4.5",
  "types-python-dateutil>=2.9.0.20250822",
]

[tool.mypy]
files = "."
strict = true
enable_error_code = [
  "deprecated",
  "explicit-override",
  "ignore-without-code",
  "redundant-expr",
  "truthy-bool",
  "truthy-iterable",
  "unused-awaitable",
]
warn_unreachable = true
plugins = [
  "mypy_django_plugin.main",
  "mypy_drf_plugin.main",
  "pydantic.mypy",
]

[[tool.mypy.overrides]]
module = [
  "channels.*",
  "geopy.*",
  "redis_entraid.*",
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "treasure.settings"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff]
extend-exclude = [
  "hunt/migrations",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "C90",               # mccabe
  "COM",               # flake8-commas
  "CPY",               # flake8-copyright
  "D",                 # pydocstyle
  "DOC",               # pydoclint
  "ERA",               # eradicate
  "PD",                # pandas-vet
  "PLR",               # pylint-refactor
  "T20",               # flake8-print
  "ANN401",            # any-type
  "E203",              # whitespace before ':'
  "S101",              # assert-used
]
extend-safe-fixes = [
  "TC",                # flake8-type-checking
]
unfixable = [
  "F841",              # unused-variable
]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-type-checking]
strict = true

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.pytest]
DJANGO_SETTINGS_MODULE = "treasure.test_settings"
python_files = ["test_*.py"]
